refactor: types util to correct directory
