+++ lib/views/components/async/loading_icon_button.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:flutter_hooks/flutter_hooks.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-import "package:proxima/utils/ui/types.dart";
+import "package:proxima/utils/types.dart";
 
 enum LoadingState {
   still,
