+++ lib/views/pages/profile/info_cards/profile_info_card.dart
@@ -1,5 +1,5 @@
 import "package:flutter/material.dart";
-import "package:proxima/utils/ui/types.dart";
+import "package:proxima/utils/types.dart";
 import "package:proxima/views/components/async/loading_icon_button.dart";
 import "package:proxima/views/pages/profile/info_cards/profile_info_pop_up.dart";
