+++ lib/views/pages/profile/info_cards/profile_info_column.dart
@@ -1,5 +1,5 @@
 import "package:flutter/material.dart";
-import "package:proxima/utils/ui/types.dart";
+import "package:proxima/utils/types.dart";
 
 /// This widget defines the style of the columns in the profile page,
 /// such as the posts and comments columns
